Create a separate api_response go file
